Add `Float64` type signatures

+++ example.jl
@@ -13,8 +13,8 @@ include("src/SampleFunctions.jl")
 # ╔═╡ 981a1eae-8801-45b8-a401-6394a0ca5248
 begin
 	λ = 0.6
-	u(c) = SampleFunctions.u(c, λ)
-	u′(c) = autodiff(Reverse, u, Active, Active(c))[1][1]
+	u(c::Float64) = SampleFunctions.u(c, λ)
+	u′(c::Float64) = autodiff(Reverse, u, Active, Active(c))[1][1]
 end
 
 # ╔═╡ 8a2682b1-24fd-41ea-a7e6-c656b8705092
